flashfs: use uintptr_t instead of uint32_t in flashfs API

+++ flight/pios/common/pios_flashfs_logfs.c
@@ -897,7 +897,7 @@ static int8_t logfs_append_to_log (struct logfs_state * logfs, uint32_t obj_id, 
  * @retval -6 if filesystem is full even after garbage collection should have freed space
  * @retval -7 if writing the new object to the filesystem failed
  */
-int32_t PIOS_FLASHFS_ObjSave(uint32_t fs_id, uint32_t obj_id, uint16_t obj_inst_id, uint8_t * obj_data, uint16_t obj_size)
+int32_t PIOS_FLASHFS_ObjSave(uintptr_t fs_id, uint32_t obj_id, uint16_t obj_inst_id, uint8_t * obj_data, uint16_t obj_size)
 {
 	int8_t rc;
 
@@ -983,7 +983,7 @@ out_exit:
  * @retval -4 if object size in filesystem does not exactly match buffer size
  * @retval -5 if reading the object data from flash fails
  */
-int32_t PIOS_FLASHFS_ObjLoad(uint32_t fs_id, uint32_t obj_id, uint16_t obj_inst_id, uint8_t * obj_data, uint16_t obj_size)
+int32_t PIOS_FLASHFS_ObjLoad(uintptr_t fs_id, uint32_t obj_id, uint16_t obj_inst_id, uint8_t * obj_data, uint16_t obj_size)
 {
 	int8_t rc;
 
@@ -1050,7 +1050,7 @@ out_exit:
  * @retval -2 if failed to start transaction
  * @retval -3 if failed to delete the object from the filesystem
  */
-int32_t PIOS_FLASHFS_ObjDelete(uint32_t fs_id, uint32_t obj_id, uint16_t obj_inst_id)
+int32_t PIOS_FLASHFS_ObjDelete(uintptr_t fs_id, uint32_t obj_id, uint16_t obj_inst_id)
 {
 	int8_t rc;
 
@@ -1091,7 +1091,7 @@ out_exit:
  * @retval -4 if failed to activate arena 0
  * @retval -5 if failed to mount arena 0
  */
-int32_t PIOS_FLASHFS_Format(uint32_t fs_id)
+int32_t PIOS_FLASHFS_Format(uintptr_t fs_id)
 {
 	int32_t rc;
